クラスを抽出
