+++ samples/web/table-authentication/src/main/java/keel/tableauth/AdminController.java
@@ -1,0 +1,13 @@
+package keel.tableauth;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class AdminController {
+
+    @GetMapping("/admin")
+    public String admin() {
+        return "/admin";
+    }
+}
