+++ samples/web/table-authentication/src/main/java/keel/tableauth/TopController.java
@@ -1,0 +1,13 @@
+package keel.tableauth;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class TopController {
+
+    @GetMapping("/top")
+    public String top() {
+        return "top";
+    }
+}
