+++ samples/web/table-authentication/src/main/java/keel/tableauth/UserController.java
@@ -1,0 +1,13 @@
+package keel.tableauth;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class UserController {
+
+    @GetMapping("/user")
+    public String user() {
+        return "/user";
+    }
+}
