+++ samples/web/table-authentication/src/main/java/keel/tableauth/WebMvcConfig.java
@@ -1,0 +1,15 @@
+package keel.tableauth;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebMvcConfig implements WebMvcConfigurer {
+
+    @Override
+    public void addViewControllers(ViewControllerRegistry registry) {
+        registry.addViewController("/login")
+                .setViewName("login");
+    }
+}
